Rename SizeSelect -> LegacySizeSelect
